Revert "Refactor and improve program exports (#133)"
This reverts commit c8bad44c54533620176fd41b366da1a3f58a4645.
