--- packages/gill/src/programs/address-lookup-table/index.ts
@@ -1,6 +1,0 @@
-/**
- * Solana Program Library (SPL) Address Lookup Table client
- * See: https://github.com/solana-program/address-lookup-table
- */
-
-export * from "@solana-program/address-lookup-table";
