--- packages/gill/src/programs/compute-budget/index.ts
@@ -1,7 +1,0 @@
-/**
- * Solana Program Library (SPL) Compute Budget Program client
- * See: https://github.com/solana-program/compute-budget
- */
-
-export * from "@solana-program/compute-budget";
-export * from "./helpers";
