+++ packages/gill/src/programs/index.ts
@@ -1,17 +1,16 @@
 /**
  * Re-export the most common program clients compatible with this library
- *
- * Note: These reexports are placed into separate directories so the
- * gill docs can be generated to include them
  */
-export * from "./address-lookup-table";
+
+export * from "@solana-program/system";
+
+export * from "@solana-program/address-lookup-table";
+export * from "@solana-program/compute-budget";
 export * from "./compute-budget";
-export * from "./system";
-export * from "./token";
 
 /**
  * Codama generated clients, stored internally in this package
  * (and associated helpers for them)
  */
-export * from "./memo"; // generated vice reexported from @solana-program/memo
+export * from "./memo"; // vendored in from @solana-program/memo
 export * from "./token-metadata";
