+++ packages/gill/src/programs/memo/index.ts
@@ -1,6 +1,1 @@
-/**
- * Solana Program Library (SPL) Memo Program client
- * See: https://github.com/solana-program/memo
- */
-
 export * from "./generated";
