--- packages/gill/src/programs/system/index.ts
@@ -1,6 +1,0 @@
-/**
- * Solana Program Library (SPL) System Program client
- * See: https://github.com/solana-program/system
- */
-
-export * from "@solana-program/system";
